Unir la clase de ImageLoader
